bug fixes (pyqt display)

+++ epyseg/draw/widgets/paint.py
@@ -57,7 +57,7 @@ class Createpaintwidget(QWidget):
         height = self.image.size().height()
         top = self.geometry().x()
         left = self.geometry().y()
-        self.setGeometry(top, left, width*self.scale, height*self.scale)
+        self.setGeometry(top, left, int(width*self.scale), int(height*self.scale))
         self.imageDraw = QtGui.QImage(self.image.size(), QtGui.QImage.Format_ARGB32)
         self.imageDraw.fill(QtCore.Qt.transparent)
         self.cursor = QtGui.QImage(self.image.size(), QtGui.QImage.Format_ARGB32)
@@ -124,8 +124,8 @@ class Createpaintwidget(QWidget):
         # draw the new one
         painter.setPen(QtGui.QPen(self.cursorColor, 2, QtCore.Qt.SolidLine, QtCore.Qt.RoundCap,
                                   QtCore.Qt.RoundJoin))
-        painter.drawEllipse(zoom_corrected_pos, self.brushSize / 2.,
-                            self.brushSize / 2.)
+        painter.drawEllipse(zoom_corrected_pos, int(self.brushSize / 2.),
+                            int(self.brushSize / 2.))
         painter.end()
         region = self.scrollArea.widget().visibleRegion()
         self.update(region)
@@ -172,8 +172,8 @@ class Createpaintwidget(QWidget):
         # the corresponding rect
         visibleRect = visibleRegion.boundingRect()
         # the visibleRect taking zoom into account
-        scaledVisibleRect = QRect(visibleRect.x() / self.scale, visibleRect.y() / self.scale,
-                                  visibleRect.width() / self.scale, visibleRect.height() / self.scale)
+        scaledVisibleRect = QRect(int(visibleRect.x() / self.scale), int(visibleRect.y() / self.scale),
+                                  int(visibleRect.width() / self.scale), int(visibleRect.height() / self.scale))
         if self.image is None:
             canvasPainter.eraseRect(visibleRect)
             canvasPainter.end()
